requirements.txt
tensorflow 
pillow 
numpy 
scikit-learn 
matplotlib 
pandas
flask
flask-cors